+++ smol/cofe/wrangling/wrangler.py
@@ -18,9 +18,11 @@ __credits__ = "William Davidson Richard"
 
 from typing import Sequence
 import warnings
+from itertools import combinations
 import numpy as np
 from monty.json import MSONable, jsanitize
 from pymatgen import Structure
+from pymatgen.analysis.structure_matcher import StructureMatcher
 from smol.cofe.space.clusterspace import ClusterSubspace
 from smol.exceptions import StructureMatchError
 
@@ -162,6 +164,22 @@ class StructureWrangler(MSONable):
         cols = cols if cols is not None else range(self.num_features)
         return np.linalg.matrix_rank(self.feature_matrix[rows][:, cols])
 
+    def get_feature_matrix_orbit_rank(self, orbit_id, rows=None):
+        """Get the rank of an orbit submatrix of the feature matrix.
+
+        Args:
+            orbit_id (int):
+                Orbit id to obtain sub feature matrix rank of.
+            rows (list): optional
+                List of row indices corresponding to structures to include.
+
+        Returns:
+            int: rank of orbit sub feature matrix
+        """
+        columns = [i for i, oid in enumerate(self._subspace.function_orbit_ids)
+                   if oid == orbit_id]
+        return self.get_feature_matrix_rank(rows=rows, cols=columns)
+
     def get_condition_number(self, rows=None, cols=None, p=2):
         """Compute the condition number for the feature matrix or submatrix.
 
@@ -212,9 +230,16 @@ class StructureWrangler(MSONable):
             X /= np.sqrt(X.T.dot(X).diagonal())
         return X.T.dot(X)
 
-    def get_duplicate_corr_inds(self):
+    def get_duplicate_corr_indices(self, decimals=12):
         """Find indices of rows with duplicate corr vectors in feature matrix.
 
+        Args:
+            decimals (int): optional
+                number of decimals to round correlations in order to allow
+                some numerical tolerance for finding duplicates. If None is
+                given no rounding will be done. Beware that orthogonal basis
+                will likeley be off by some numerical tolerance so rounding is
+                recommended.
         Returns:
             list: list containing lists of indices of rows in feature_matrix
             where duplicates occur
@@ -224,12 +249,57 @@ class StructureWrangler(MSONable):
         else:
             num_ext = len(self.cluster_subspace.external_terms)
             end = self.feature_matrix.shape[1] - num_ext - 1
-            _, inverse = np.unique(self.feature_matrix[:, :end],
+            feature_matrix = self.feature_matrix if decimals is None \
+                else np.around(self.feature_matrix, decimals,
+                               self.feature_matrix.copy())
+            _, inverse = np.unique(feature_matrix[:, :end],
                                    return_inverse=True, axis=0)
             duplicate_inds = [list(np.where(inverse == i)[0])
                               for i in np.unique(inverse)
                               if len(np.where(inverse == i)[0]) > 1]
         return duplicate_inds
+
+    def get_matching_corr_duplicate_indices(self, decimals=12,
+                                            structure_matcher=None,
+                                            **matcher_kwargs):
+        """Find indices of equivalent structures.
+
+        Args:
+            decimals (int): optional
+                number of decimals to round correlations in order to allow
+                some numerical tolerance for finding duplicates.
+            structure_matcher (StructureMatcher): optional
+                A StructureMatcher object to use for matching structures.
+            **matcher_kwargs:
+                Keyword arguments to use when initializing a structure matcher
+                if not given
+
+        Returns:
+            list: list of lists of equivalent structures (that match) and have
+                  duplicate correlation vectors.
+        """
+        matcher = structure_matcher if structure_matcher is not None \
+            else StructureMatcher(**matcher_kwargs)
+        duplicate_indices = self.get_duplicate_corr_indices(decimals)
+
+        matching_inds = []
+        for inds in duplicate_indices:
+            # match all combinations of duplicates
+            matches = [set(c) for c in combinations(inds, 2) if
+                       matcher.fit(self.structures[c[0]],
+                                   self.structures[c[1]],
+                                   symmetric=True)]
+            while overlaps := list(
+                    filter(lambda s: s[0] & s[1], combinations(matches, 2))):
+                all_overlaps = [o for overlap in overlaps for o in overlap]
+                # keep only disjoint sets
+                matches = [s for s in matches if s not in all_overlaps]
+                # add union of overlapping sets
+                for s1, s2 in overlaps:
+                    if s1 | s2 not in matches:
+                        matches.append(s1 | s2)
+            matching_inds += [list(sorted(m)) for m in matches]
+        return matching_inds
 
     def get_constant_features(self):
         """Find indices of constant feature vectors (columns).
@@ -362,11 +432,17 @@ class StructureWrangler(MSONable):
         """
         keys = ['structure', 'ref_structure', 'properties', 'weights',
                 'scmatrix', 'mapping', 'features', 'size']
-        for item in data_items:
+        for i, item in enumerate(data_items):
             if not all(key in keys for key in item.keys()):
                 raise ValueError(
-                    "A supplied data item is missing required keys. Make sure"
+                    f"Data item {i} is missing required keys. Make sure"
                     " they were obtained with the process_structure method.")
+            if len(self._items) > 0:
+                if not all(prop in self._items[0]['properties'].keys()
+                           for prop in item['properties'].keys()):
+                    raise ValueError(
+                        f"Data item {i} is missing one of the following "
+                        f"properties: {self.available_properties}")
             self._items.append(item)
             self._corr_duplicate_warning(self.num_structures - 1)
 
@@ -409,7 +485,8 @@ class StructureWrangler(MSONable):
                 "Length of property_vector must match number of structures"
                 f" {len(property_vector)} != {self.num_structures}.")
         if normalized:
-            property_vector *= self.sizes
+            # make copy
+            property_vector = self.sizes * property_vector.copy()
 
         for prop, item in zip(property_vector, self._items):
             item['properties'][key] = prop
@@ -551,7 +628,7 @@ class StructureWrangler(MSONable):
 
     def _corr_duplicate_warning(self, index):
         """Warn if corr vector of item with given index is duplicated."""
-        for duplicate_inds in self.get_duplicate_corr_inds():
+        for duplicate_inds in self.get_duplicate_corr_indices():
             if index in duplicate_inds:
                 duplicates = "".join(
                     f"Index {i} - {self._items[i]['structure'].composition}"